+++ core/database/src/main/kotlin/ir/vmessenger/core/database/VMessengerDatabase.kt
@@ -20,6 +20,7 @@ import ir.vmessenger.core.database.dao.LocationAccessDao
 import ir.vmessenger.core.database.dao.LocationSampleDao
 import ir.vmessenger.core.database.dao.LocationShareDao
 import ir.vmessenger.core.database.dao.MailboxDao
+import ir.vmessenger.core.database.dao.ManagedNodeDao
 import ir.vmessenger.core.database.dao.MessageDao
 import ir.vmessenger.core.database.dao.MessageEditHistoryDao
 import ir.vmessenger.core.database.dao.MessageRecipientDao
@@ -42,6 +43,7 @@ import ir.vmessenger.core.database.entity.LocationAccessEntity
 import ir.vmessenger.core.database.entity.LocationSampleEntity
 import ir.vmessenger.core.database.entity.LocationShareEntity
 import ir.vmessenger.core.database.entity.MailboxBlobEntity
+import ir.vmessenger.core.database.entity.ManagedNodeEntity
 import ir.vmessenger.core.database.entity.MessageEditHistoryEntity
 import ir.vmessenger.core.database.entity.MessageEntity
 import ir.vmessenger.core.database.entity.MessageRecipientEntity
@@ -72,6 +74,7 @@ import ir.vmessenger.core.database.entity.RelayNodeEntity
         LocationSampleEntity::class,
         MailboxBlobEntity::class,
         DhtRecordEntity::class,
+        ManagedNodeEntity::class,
     ],
     version = DatabaseSchema.VERSION,
     exportSchema = true,
@@ -101,4 +104,5 @@ abstract class VMessengerDatabase : RoomDatabase() {
     abstract fun locationSampleDao(): LocationSampleDao
     abstract fun mailboxDao(): MailboxDao
     abstract fun dhtRecordDao(): DhtRecordDao
+    abstract fun managedNodeDao(): ManagedNodeDao
 }